+++ packages/react-html/src/components/tests/Meta.test.tsx
@@ -15,4 +15,47 @@ describe('<Meta />', () => {
 
     expect(spy).toHaveBeenCalledWith(props);
   });
+
+  it('only keeps the last added meta based on name or property', () => {
+    const globalDescription = 'global description';
+    const pageDescription = 'page description';
+
+    const manager = new HtmlManager();
+
+    mountWithManager(
+      <>
+        <Meta name="desciption" content={globalDescription} />
+        <Meta property="og:description" content={globalDescription} />
+        <Meta name="desciption" content={pageDescription} />
+        <Meta property="og:description" content={pageDescription} />
+      </>,
+      manager,
+    );
+
+    expect(manager.state.metas).toMatchObject([
+      {
+        content: 'page description',
+        property: 'og:description',
+      },
+      {
+        content: 'page description',
+        name: 'desciption',
+      },
+    ]);
+  });
+
+  it('removes meta after unmounting', () => {
+    const manager = new HtmlManager();
+
+    const meta = mountWithManager(
+      <Meta name="desciption" content="test" />,
+      manager,
+    );
+
+    expect(manager.state.metas).toHaveLength(1);
+
+    meta.unmount();
+
+    expect(manager.state.metas).toHaveLength(0);
+  });
 });